refactor directory tree
