+++ src/engine2/common/sm/desktop_sm.py
