+++ src/engine2/common/sm/template_sm.py
